requirements.txt
llama-index-core>=0.10.40
llama-index-embeddings-openai>=0.1.10
llama-index-llms-openai>=0.1.20
llama-index-readers-file>=0.1.25
llama-index-postprocessors-rankgpt-rerank>=0.1.3
openai>=1.40.0
streamlit>=1.36.0
pandas>=2.0.0
openpyxl>=3.1.0
